test size_t

+++ sort.h
@@ -31,6 +31,6 @@ void top_down_merge(int *array, int begin, int end, int *copy);
 void top_down_split_merge(int *array, int begin, int end, int *copy);
 void swap1(listint_t *tprev, listint_t *tnext);
 void swap2(listint_t *tnext, listint_t *tprev);
-void quick_sort(int *array, int size);
+void quick_sort(int *array, size_t size);
 void swap3(listint_t **list, listint_t *tprev, listint_t *tnext);
 #endif
